+++ modules/elementary_functions/src/cpp/IsNan.cpp
@@ -23,9 +23,10 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
+#include "nlsConfig.h"
+#include <Eigen/Dense>
 #include "IsNaN.hpp"
 #include "ClassName.hpp"
-#include <Eigen/Dense>
 //=============================================================================
 namespace Nelson {
 //=============================================================================
@@ -33,7 +34,10 @@ template <class T>
 void
 boolean_isnan(indexType N, logical* C, const T* A)
 {
-    for (indexType i = 0; i < N; i++) {
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+    for (ompIndexType i = 0; i < (ompIndexType)N; i++) {
         C[i] = std::isnan(A[i]);
     }
 }
@@ -42,7 +46,10 @@ template <class T>
 void
 boolean_isnan_cplx(indexType N, logical* C, const T* A)
 {
-    for (indexType i = 0; i < N; i++) {
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+    for (ompIndexType i = 0; i < (ompIndexType)N; i++) {
         C[i] = std::isnan(A[i].real()) || std::isnan(A[i].imag());
     }
 }
@@ -99,7 +106,11 @@ IsNaN(ArrayOf A)
         void* Cp
             = Nelson::ArrayOf::allocateArrayOf(NLS_LOGICAL, A.getLength(), stringVector(), false);
         auto* CpLogical = static_cast<logical*>(Cp);
-        for (indexType i = 0; i < A.getLength(); i++) {
+        ompIndexType N = (ompIndexType)A.getLength();
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+        for (ompIndexType i = 0; i < N; i++) {
             CpLogical[i] = static_cast<logical>(0);
         }
         C.setDataPointer(Cp);