+++ CHANGELOG.md
@@ -12,9 +12,15 @@
 
 ## Bug Fixes:
 
+- [#364](http://github.com/Nelson-numerical-software/nelson/issues/364): `isinf`, `isnan`, `conj`, `double`, `single`, `real`, `imag` are faster.
+
 - [#361](http://github.com/Nelson-numerical-software/nelson/issues/361): `abs` is faster.
 
 - [#360](http://github.com/Nelson-numerical-software/nelson/issues/360): `ctranpose` and `transpose` are faster.
+
+## Compilation:
+
+- Qt 5.15.2 on Windows (AppVeyor CI).
 
 
 # 0.4.11 (2020-11-24)